Return std::int64_t from the queue holder thread counters
The scheduler counter interface is std::int64_t at both ends.
scheduler_base declares get_queue_length and get_thread_count virtual
std::int64_t, thread_pool_base's whole get_thread_count_* family is
std::int64_t, local_queue_scheduler returns std::int64_t, and
thread_queue_mc::get_queue_length_staged/_pending return std::int64_t.

queue_holder_thread and queue_holder_numa sit between those two ends and
return std::size_t, so every one of them narrows a std::int64_t on the
way through and widens it again for the caller. MSVC reports the
narrowing as C4244 on a 32 bit target; 53 of the 1586 warnings the 32
bit Windows job emits come from these two headers.

Return std::int64_t from them as well, so the value keeps its type all
the way from thread_queue_mc up to shared_priority_queue_scheduler.
No conversion is left to silence and no cast is introduced.

The only consumers are in shared_priority_queue_scheduler, which already
accumulates into std::int64_t or compares against zero, so nothing at
the call sites changes.

Refs #7576

+++ libs/core/schedulers/include/hpx/schedulers/queue_holder_numa.hpp
@@ -209,11 +209,11 @@ namespace hpx::threads::policies {
         }
 
         // ----------------------------------------------------------------
-        inline std::size_t get_thread_count(
+        inline std::int64_t get_thread_count(
             thread_schedule_state state = thread_schedule_state::unknown,
             thread_priority priority = thread_priority::default_) const
         {
-            std::size_t len = 0;
+            std::int64_t len = 0;
             for (auto& q : queues_)
                 len += q->get_thread_count(state, priority);
             return len;
